core/exec-credential: simplify credential dir installation logic
Currently, if the final place is a mount but empty, we just
carry on and acquire creds in it. This is a bit weird though,
especially after the previous commit, which causes cred dir
installation to be skipped already when empty.

To make things more straightforward, let's simplify the logic
so that if final place is mounted, we either dump it and start
afresh or just reuse it no matter what.