wip: reload creds with RefreshCredentialsOnReload= or so
Need to find out how to handle the mount in the root
mountns. In the ns we definitely want to use
bind_mount_in_namespace, but what about the final
vs workspace stuff? Especially now the final dir is a
mess

!!! Follow the restrictions in unit_verify_contexts()
!!! also in the actual impl

There can be three ways in which the cred dir is set up.
1. tmpfs + service uses mountns
2. tmpfs + root ns
3. plain dir

3. should definitely be refused. 2. means we need to atomically
replace cred mount in root ns too.

If the cred dir is empty, we should be able to still reuse it.